--- PythonToolbox/user_data.json
@@ -1,7 +1,0 @@
-{
-    "authorName": "QuantConnect User",
-    "authorPicture": "AuthorProfile.png",
-    "authorBiography": "Put your biography here.",
-    "projectName": "Basic Template Algorithm",
-    "projectDescription": "Basic Template Algorithm"
-}
